+++ src/geo/coordinatesresolver.h
@@ -31,7 +31,7 @@ inline bool operator==(const PlaceInfo &lhs, const PlaceInfo &rhs)
     return lhs.country == rhs.country && lhs.town == rhs.town && lhs.location == rhs.location;
 }
 
-inline uint qHash(const PlaceInfo &key, uint seed = 0)
+inline size_t qHash(const PlaceInfo &key, size_t seed = 0)
 {
     return qHashMulti(seed, key.country, key.town, key.location.latitude(), key.location.longitude()/*,
                       key.message*/);
